+++ extensions/browsers-profiles/src/lib/firefox.ts
@@ -1,6 +1,7 @@
 import fs from "fs";
 import os from "os";
 import ini from "ini";
+import { join } from "path";
 
 import browsers from "./supported-browsers.json";
 
@@ -22,14 +23,14 @@ export const getFirefoxProfiles = () => {
   const profiles: FirefoxProfiles[] = [];
 
   browsers.firefox.forEach((browser) => {
-    const path = `${os.homedir()}${browser.path}`;
+    const path = join(os.homedir(), browser.path, "profiles.ini");
     const exists = fs.existsSync(path);
 
     if (!exists) {
       return null;
     }
 
-    const file = fs.readFileSync(`${path}/profiles.ini`, "utf-8");
+    const file = fs.readFileSync(path, "utf-8");
     const config = ini.parse(file);
 
     const browserProfiles: FirefoxProfile[] = [];