+++ extensions/manotori/src/types.ts
@@ -39,8 +39,32 @@ export type DomainEvent = {
   created_at: string;
 };
 
+export type NameserverSet = {
+  nameserver_set_id: number;
+  provider_id: number | null;
+  name: string;
+  type: "INTERNAL" | "EXTERNAL";
+  nameservers: Array<{
+    nameserver: string;
+  }>;
+  is_editable: 0 | 1;
+  is_deletable: 0 | 1;
+  is_deprecated: 0 | 1;
+};
+
+export type ErrorResponse = {
+  data:
+    | []
+    | {
+        code: number;
+        file: string;
+        line: number;
+      };
+  msg: string;
+  msg_key: null;
+};
 export type SuccessResult<T> = {
   data: T;
-  msg: string;
+  msg: string | null;
   msg_key: null;
 };