+++ extensions/safari/src/utils.ts
@@ -5,6 +5,7 @@ import osascript from "osascript-tag";
 import { URL } from "url";
 import { langAdaptor, PinyinHandler } from "./lang-adaptor";
 import { HistoryItem, LooseTab } from "./types";
+import { runAppleScript } from "@raycast/utils";
 
 type Preferences = {
   safariAppIdentifier: string;
@@ -124,3 +125,11 @@ export const groupHistoryByDay = (groups: Map<string, HistoryItem[]>, entry: His
   groups.set(date, group);
   return groups;
 };
+
+export async function getCurrentTabName() {
+  return await runAppleScript(`tell application "${safariAppIdentifier}" to return name of front document`);
+}
+
+export async function getCurrentTabURL() {
+  return await runAppleScript(`tell application "${safariAppIdentifier}" to return URL of front document`);
+}