+++ extensions/browsers-profiles/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Raycast Browsers Profiles
 
+## [Fix Firefox] - 2025-01-07
+
+- Fix extensions error when `profiles.ini` is missing from the Firefox folder 
+
 ## [Developer Edition] - 2024-01-29
 - Added Developer Editions of Chrome, Edge, and Firefox
 - Updated and added missing icons